refactor(frontend): extract my courses derived hook
